% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_all_labels.R
\name{apply_all_labels}
\alias{apply_all_labels}
\title{Apply all labels}
\usage{
apply_all_labels(lst)
}
\arguments{
\item{lst}{List of data frames containing study data}
}
\description{
Convert all numeric variables to factors based on
meta-data available in the package look up table
}
