% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_labels.R
\name{apply_labels}
\alias{apply_labels}
\title{Apply labels}
\usage{
apply_labels(df, variable, lut = internal$lut, convert_to_factor = TRUE)
}
\arguments{
\item{df}{data frame}

\item{variable}{character string representing the name of the target
variable in \code{df}}

\item{lut}{lookup table}

\item{convert_to_factor}{boolean to flat whether to conver the returned
variable to an unordered factor - defaults to TRUE}
}
\value{
character vector where NA is introduced for unmatched values
}
\description{
Apply text labels to variables which are represented in the
internal package lookup table
}
