% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_baseline_smear.R
\name{calculate_baseline_smear}
\alias{calculate_baseline_smear}
\title{Calculate baseline smear result}
\usage{
calculate_baseline_smear(baseline, myco)
}
\arguments{
\item{baseline}{data frame containing unique subject ID and their
treatment initiation date}

\item{myco}{data frame containing a subject ID, test type, date of
speciment collection and the test result}
}
\value{
data frame with baseline data and additional smear factor variable
}
\description{
Calculate baseline smear result
}
