% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_conversion.R
\name{calculate_conversion}
\alias{calculate_conversion}
\title{Calculate conversion date}
\usage{
calculate_conversion(subject_df, tolerance = 30L)
}
\arguments{
\item{subject_df}{data frame with ID, date and result variables}

\item{tolerance}{integer to represent number of days between negative
results which constitute conversion}
}
\value{
data frame with subject ID and conversion date variable
}
\description{
Calculate conversion date
}
