% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_eos_outcome.R
\name{calculate_eos_outcome}
\alias{calculate_eos_outcome}
\title{Calculate End of Study outcome}
\usage{
calculate_eos_outcome(df)
}
\arguments{
\item{df}{data frame for one subject with end of
treatment and follow up outcomes with their
respective date variables}
}
\value{
data frame with
}
\description{
Follow up status is collected longitudinally,
so the first unsuccessful outcome with its date
must be retrieved
}
