% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_final_follow_up.R
\name{calculate_final_follow_up}
\alias{calculate_final_follow_up}
\title{Calculate final follow up}
\usage{
calculate_final_follow_up(df, eval_months = c(3, 6, 9, 12))
}
\arguments{
\item{df}{data frame of all participants' baseline characteristics}

\item{eval_months}{evaluation months}
}
\description{
Calculate the final follow up date, month and status for
participants who successfully completed treatment
}
