% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_follow_up_days.R
\name{calculate_follow_up_days}
\alias{calculate_follow_up_days}
\title{Calculate follow-up time in days}
\usage{
calculate_follow_up_days(df)
}
\arguments{
\item{df}{data frame with baseline characteristics}
}
\value{
datafrme with baseline characterists and fu_days
numerical variable
}
\description{
Calculate follow-up time in days
}
