% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_lr_statistic.R
\name{calculate_lr_statistic}
\alias{calculate_lr_statistic}
\title{Calculate Likelihood Ratio statistic}
\usage{
calculate_lr_statistic(model, df)
}
\arguments{
\item{model}{Cox PH model}

\item{df}{data frame of input daata}
}
\value{
named list of p-values - one for each predictor
}
\description{
Calculate the LRT p-value for each predictor variable in a
Cox PH model
}
