% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_result_diffs.R
\name{calculate_result_diffs}
\alias{calculate_result_diffs}
\title{Calculate differences in results by days}
\usage{
calculate_result_diffs(
  results,
  date_var = "date",
  result_var = "result",
  positive_results = list(1, "MTB complex"),
  negative_results = list(3, "No growh")
)
}
\arguments{
\item{results}{participant results data frame}

\item{date_var}{string - name of date variable}

\item{result_var}{string - name of result variable}

\item{positive_results}{list of results which reset the difference
vector to zero}

\item{negative_results}{list of results which continue the difference
vector}
}
\value{
vector of doubles with with NA representing a positive result
}
\description{
Calculate differences in results by days
}
