% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_variables.R
\name{calculate_variables}
\alias{calculate_variables}
\title{Calculate study variables}
\usage{
calculate_variables(df)
}
\arguments{
\item{df}{data frame of baseline characteristics}
}
\value{
data frame of baseline characteritcs with additional
variables
}
\description{
Calculate study variables
}
