% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_factors.R
\name{combine_ae_grades}
\alias{combine_ae_grades}
\title{Combine AE grades}
\usage{
combine_ae_grades(df, target, result)
}
\arguments{
\item{df}{data frame of baseline characteristics}

\item{target}{string representing the variable being targetted}

\item{result}{string name of the resulting variable}
}
\description{
A custom function to combine AE variables with multiple grades into
a binary varaible
}
