% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cc_days.R
\name{create_cc_days}
\alias{create_cc_days}
\title{Calculate time to culture conversion}
\usage{
create_cc_days(trtstdat, convdat)
}
\arguments{
\item{trtstdat}{date vector representing the treatment initation date}

\item{convdat}{date vector representing the study team's recorded culture
conversion date.}
}
\value{
numeric vector representing the number of days to culture conversion
}
\description{
The mSTR study data contains information about each sputum result
and the study team's calculation for culture conversion.
This function compares the treatment start date with the study team's
recorded culture conversion date to generate a numerical result representing
the number of days until culture conversion.
}
