% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_conversion_variables.R
\name{create_conversion_variables}
\alias{create_conversion_variables}
\title{Create conversion variables}
\usage{
create_conversion_variables(baseline, myco)
}
\arguments{
\item{baseline}{data frame of baseline participant characteristics}

\item{myco}{data frame of participant mycobacteriology results}
}
\value{
data frame with globalrecordid, cc_days and
cc_event variables
}
\description{
This function generates conversion variables including
a time at risk and an event status
}
