% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{diff_days}
\alias{diff_days}
\title{Calculate days between two dates}
\usage{
diff_days(start, end)
}
\arguments{
\item{start}{start date}

\item{end}{end date}
}
\description{
Calculate days between two dates
}
