% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_p_values}
\alias{format_p_values}
\title{Format p-values for result tables}
\usage{
format_p_values(x)
}
\arguments{
\item{x}{vector of numeric p-values}
}
\description{
The format has been aligned with the request from the Lancet
ID editors.
}
