% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_factors.R
\name{handle_factors}
\alias{handle_factors}
\title{Handle factors}
\usage{
handle_factors(baseline_df)
}
\arguments{
\item{baseline_df}{baseline characteristics data frame}
}
\description{
Clean, relevel, and generally prepare factor variables
in the baseline data frame for analysis
}
