% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_positive_baseline_culture.R
\name{has_positive_baseline_culture}
\alias{has_positive_baseline_culture}
\title{Has positive baseline culture}
\usage{
has_positive_baseline_culture(
  baseline,
  myco,
  culture_type = "culq",
  var_suffix = NULL
)
}
\arguments{
\item{baseline}{data frame of baseline participant characteristics}

\item{myco}{data frame of participant mycobacteriology results}

\item{culture_type}{string vector to define valid culture types}

\item{var_suffix}{string suffix for returned variable}
}
\description{
Has positive baseline culture
}
