% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_ae_grade.R
\name{match_ae_grade}
\alias{match_ae_grade}
\title{Match AE grades to values}
\usage{
match_ae_grade(values, grades)
}
\arguments{
\item{values}{vector of numerical values}

\item{grades}{vector of numbers representing the left-boundary of grading
levels}
}
\description{
Match AE grades to values
}
