% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_and_filter_by_start_date}
\alias{merge_and_filter_by_start_date}
\title{Merge and Filter by start date}
\usage{
merge_and_filter_by_start_date(baseline, target, target_date_var)
}
\arguments{
\item{baseline}{Data frame of baseline study data.}

\item{target}{Data frame of target study data.}

\item{target_date_var}{String representing the date variable in the target.}
}
\value{
Data frame
}
\description{
Merge a data frame with the \code{baseline} data frame then filter entries
according to the treatment initaition date
}
