% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_imported.R
\name{merge_imported}
\alias{merge_imported}
\title{Merged imported study data}
\usage{
merge_imported(df_list, multi_country = FALSE)
}
\arguments{
\item{df_list}{list of named dataframes returned from
\code{read_in}}

\item{multi_country}{Boolean to allow data from multiple
countries to be combined}
}
\value{
Named list of data frames
}
\description{
Some study teams will have multiple data files
which require merging prior to analysis. This function
takes the nested list created by \code{read_in} and merges
it.
}
