% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ae_summary_data.R
\name{prepare_ae_summary_data}
\alias{prepare_ae_summary_data}
\title{Prepare AE summary data}
\usage{
prepare_ae_summary_data(baseline, adverse, type = c("SAE", "AE"))
}
\arguments{
\item{baseline}{data frame representing the baseline study data}

\item{adverse}{data frame representing adverse event data}

\item{type}{SAE or AE represents the type of AE required}
}
\description{
Prepare AE summary data
}
