% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_baseline.R
\name{prepare_baseline}
\alias{prepare_baseline}
\title{Prepare baseline data frame}
\usage{
prepare_baseline(df_list)
}
\arguments{
\item{df_list}{List of named data frames}
}
\description{
Prepare baseline data frame
}
