% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_eos_outcomes.R
\name{prepare_eos_outcomes}
\alias{prepare_eos_outcomes}
\title{Reshape EoS outcome variables}
\usage{
prepare_eos_outcomes(df)
}
\arguments{
\item{df}{data frame containing unique ID, EoS outcomes and
EoS evaluation dates}
}
\description{
Reshape EoS outcome variables
}
