% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_myco.R
\name{prepare_myco}
\alias{prepare_myco}
\title{Prepare myco data for analysis}
\usage{
prepare_myco(myco, baseline, purpose = c("quality"))
}
\arguments{
\item{myco}{Data frame of study mycobacterial data.}

\item{baseline}{Data frame of baseline study data.}

\item{purpose}{String variable to represent the puprpose of data preparation
Allowed values are:
* quality}
}
\value{
A data frame merged between \code{baseline} and \code{myco} with selected
variables
}
\description{
Prepare myco data for analysis
}
