% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_quality_data.R
\name{prepare_quality_data}
\alias{prepare_quality_data}
\title{Prepare quality report data}
\usage{
prepare_quality_data(study_data)
}
\arguments{
\item{study_data}{A named list of data frames with the following keys:
\itemize{
\item adverse
\item baseline
\item dst
\item myco
}}
}
\value{
Named list
}
\description{
Merge, condense and simplify a full set of study data into the data
structure required for the study data quality report template.
}
