% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in.R
\name{read_in}
\alias{read_in}
\title{Read in study data}
\usage{
read_in(dir_path, file_names = c("baseline", "myco"))
}
\arguments{
\item{dir_path}{directory path to study data files}

\item{file_names}{vector of file names to read in}
}
\value{
list of data frames
}
\description{
Read in study data which has been saved as CSV files
}
