% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ready.R
\name{ready}
\alias{ready}
\title{Package ready?}
\usage{
ready()
}
\value{
NULL - prints a message to the console
}
\description{
A simple function to show that the user has successfully installed
the package and are ready to use it for analyses.
}
