% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{relevel_variables}
\alias{relevel_variables}
\title{Relevel variables}
\usage{
relevel_variables(df, config)
}
\arguments{
\item{df}{data frame}

\item{config}{list}
}
\description{
Choose base level for variables
}
