% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_invalid_records.R
\name{remove_invalid_records}
\alias{remove_invalid_records}
\title{Remove user-defined invalid records}
\usage{
remove_invalid_records(df)
}
\arguments{
\item{df}{Data frame with the recstatus variable defined.}
}
\value{
Data frame with invalid records removed
}
\description{
During data entry, users can mark a record as invalid instead of deleting
it. This function removes these invalid records and logs how many rows
have been removed
}
