% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_withdrawn_subjects.R
\name{remove_withdrawn_subjects}
\alias{remove_withdrawn_subjects}
\title{Remove withdrawn subjects}
\usage{
remove_withdrawn_subjects(df)
}
\arguments{
\item{df}{baseline data frame}
}
\value{
data frame with withdrawn subject rows removed
}
\description{
Remove withdrawn subjects
}
