% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_internal_rmd.R
\name{render_internal_rmd}
\alias{render_internal_rmd}
\title{Render an internal RMarkdown file}
\usage{
render_internal_rmd(
  input_file,
  output_file,
  output_format = "html_document",
  ...
)
}
\arguments{
\item{input_file}{Path to the RMarkdown file relative to the inst/rmarkdown/ directory}

\item{output_file}{Path to the output file}

\item{output_format}{Output format for the rendered document}

\item{...}{Additional arguments to pass to the \code{rmarkdown::render()} function}
}
\description{
This function renders an internal RMarkdown file located in the inst/rmarkdown/ directory
}
