% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ae_summary.R
\name{run_ae_summary}
\alias{run_ae_summary}
\title{Create Adverse Event summary table}
\usage{
run_ae_summary(data_path, output_file)
}
\arguments{
\item{data_path}{path to data directory}

\item{output_file}{path and name of output file - must be .html}
}
\description{
Create Adverse Event summary table
}
