% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_analysis.R
\name{run_analysis}
\alias{run_analysis}
\title{Run analysis}
\usage{
run_analysis(data_path, output_dir)
}
\arguments{
\item{data_path}{path to data directory}

\item{output_dir}{path to output directory}
}
\value{
NULL - creates files in output directory
}
\description{
Run analysis
}
