% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_quality_check.R
\name{run_quality_check}
\alias{run_quality_check}
\title{Run data quality check}
\usage{
run_quality_check(data_path, output_file)
}
\arguments{
\item{data_path}{path to data directory}

\item{output_file}{path and name of output file - must be .html}
}
\value{
NULL - creates a HTML file as a side-effect
}
\description{
Run data quality check
}
