% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tutorial.R
\name{run_tutorial}
\alias{run_tutorial}
\title{Run internal tutorial}
\usage{
run_tutorial(name, ...)
}
\arguments{
\item{name}{a single word defining the name of the tutorial to be loaded}

\item{...}{optional arguments to pass to learnr::run_tutorial}
}
\description{
Run internal tutorial
}
