% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_to_date.R
\name{transform_to_date}
\alias{transform_to_date}
\title{Transform character to date}
\usage{
transform_to_date(v)
}
\arguments{
\item{v}{A character vector which represents dates}
}
\value{
A date vector
}
\description{
Some versions of the user's study data might include date variables
which are transcribed as character vectors. This function will convert
these to date variables
}
