% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_eos_outcome.R
\name{create_eos_outcome}
\alias{create_eos_outcome}
\title{Create End of Study outcome}
\usage{
create_eos_outcome(eot, eof)
}
\arguments{
\item{eot}{vector of End of treatment outcomes}

\item{eof}{vector of End of follow-up outcomes}
}
\value{
vector (factor) with the same length as the input arguments which
represents the end of study outcome
}
\description{
The original end of study outcome - \code{stat12} only provides information
about a participants status 12 months after treatment completion. It does
not incorporate any information about their on-treatment experience - for
example, if a participant fails on treatment, their \code{status12} will be
missing. This function combines the end of treatment outcome and the
\code{status12} variable to create a valid end of study outcome.
}
