% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_eos_outcome.R
\name{create_eos_outcome}
\alias{create_eos_outcome}
\title{Create End of Study outcome}
\usage{
create_eos_outcome(df)
}
\arguments{
\item{df}{baseline data frame}
}
\value{
Original baseline data frame with added variables - eos_outcome &is  eos_date
}
\description{
The original end of study outcome - \code{stat12} only provides information
about a participants status 12 months after treatment completion. It does
not incorporate any information about their on-treatment experience - for
example, if a participant fails on treatment, their \code{status12} will be
missing. This function combines the end of treatment outcome and the
\code{status12} variable to create a valid end of study outcome.
}
