% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in.R
\name{read_in}
\alias{read_in}
\title{Read in study data}
\usage{
read_in(dir_path, file_names)
}
\arguments{
\item{dir_path}{Directory path to study data files}

\item{file_names}{Named list of file names to read in.
Allowed file names are:
\itemize{
\item adverse
\item baseline
\item dst
\item myco
}}
}
\value{
A named list of data frames where the names correspond to
the names of elements provided in the \strong{file_names} parameter
}
\description{
Read in study data which has been saved as CSV files
}
